requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
deepeval==2.0.9
Deprecated==1.2.15
distro==1.9.0
docx2txt==0.8
elastic-transport==8.15.1
elasticsearch==8.17.0
execnet==2.1.1
frozenlist==1.5.0
googleapis-common-protos==1.66.0
grpcio==1.60.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.4.0
iniconfig==2.0.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langsmith==0.2.6
markdown-it-py==3.0.0
marshmallow==3.23.2
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.1
openai==1.58.1
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.10.12
packaging==24.2
pluggy==1.5.0
portalocker==3.0.0
propcache==0.2.1
protobuf==4.25.5
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.18.0
pytest==8.3.4
pytest-repeat==0.9.3
pytest-xdist==3.6.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
sentry-sdk==2.19.2
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
